+++ scripts/experiment.py
@@ -346,7 +346,7 @@ class Experiment(ABC):
         logger.info(f"Loaded datasets with sample counts: (train, val, test) = ({len(ds_train)}, {len(ds_val)}, {len(ds_test)}).")
 
         logger.info(f"Loading model: {args.model}")
-        model, tokenizer = load_model(args.model, dtype=torch.bfloat16, device_map="cuda:0")
+        model, tokenizer = load_model(args.model, device_map="cuda:0")
         targeted_model = TargetedModel(model=model, tokenizer=tokenizer, is_chat=is_chat)
         logger.info(f"\nModel architecture: {targeted_model.model}\n")
